Conflicts:
	Assets/Basic Dungeon.unity
	Assets/Resources/ef tex/ef_badak_blood.mat

+++ Assets/Script/Creature.cs
@@ -451,7 +451,7 @@ public class Creature : MonoBehaviour {
 			}
 		}
 
-		float dmg = (damageDesc.Damage+damageDesc.Damage*criticalDamage)-m_creatureProperty.PhysicalDefencePoint;
+		int dmg = (int)((damageDesc.Damage+damageDesc.Damage*criticalDamage)-m_creatureProperty.PhysicalDefencePoint);
 		dmg = Mathf.Max(0, Mathf.FloorToInt(dmg));
 		if (dmg == 0)
 		{
